Move WatchlistEntry struct to own file
